ds-measurer: feature, change, refactor. Add guides, modes of selection, changes in controls.
Introduced three modes of selection with different approaches to how to take
into account the line thickness of the cursor or selection rectangle.
Added a new object type — guides. The new type is structurally similar to the selection
object type. We can place them, switch between them, move, delete, copy their values.
Big changes in controls. More roles assigned to the mouse: now the right button
is responsible for moving the objects, the wheel for switching between them.
Also, a lot of changes in the configuration file.
Also performed a significant refactoring because of all these features and changes.