ds-measurer: feature, refactor. Masking selections, changing their colors and marks.
Added the ability to shade the area outside the selection by pressing Alt key
and moving a mouse up and down. Selections keep their masking states. When the
masking value of the selection is greater than zero: neither selection rectangle
nor cursor inside the selected area are not displayed; it is possible to hide
and show all the the marks of the selection at once by pressing a key.
Now it is also possible to change colors and marks of each selection and guide.